backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
firebase-admin==6.2.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.23
gunicorn==21.2.0
pytest==7.4.3